requirements.txt
python-dotenv==1.0.0
requests==2.28.1
datetime==4.3
tweepy==4.12.1
requests-oauth1